+++ packages/article-extras/__tests__/shared.web.js
@@ -8,24 +8,11 @@ import {
   print
 } from "@times-components/jest-serializer";
 import { iterator } from "@times-components/test-utils";
-import { AlgoliaSearchProvider } from "@times-components/utils";
+import { AlgoliaSearchProvider } from "@times-components/ts-components";
 
 import { UserState } from "./mocks";
 import ArticleExtras from "../src/article-extras";
 import { relatedArticleSlice, topics } from "../fixtures/article-extras";
-
-jest.mock("@times-components/utils", () => {
-  const actualUtils = jest.requireActual("@times-components/utils");
-  return {
-    ...actualUtils,
-    useAlgoliaSearch: () => ({
-      getRelatedArticles: () => ({
-        sliceName: "StandardSlice",
-        items: []
-      })
-    })
-  };
-});
 
 const algoliaSearchKeys = {
   applicationId: "",
@@ -59,6 +46,7 @@ export default () => {
           <AlgoliaSearchProvider
             article={{ id: "dummy-article-id" }}
             algoliaSearchKeys={algoliaSearchKeys}
+            analyticsStream={jest.fn()}
           >
             <ArticleExtras
               analyticsStream={() => {}}
@@ -85,6 +73,7 @@ export default () => {
           <AlgoliaSearchProvider
             article={{ id: "dummy-article-id" }}
             algoliaSearchKeys={algoliaSearchKeys}
+            analyticsStream={jest.fn()}
           >
             <ArticleExtras
               analyticsStream={() => {}}
@@ -110,6 +99,7 @@ export default () => {
           <AlgoliaSearchProvider
             article={{ id: "dummy-article-id" }}
             algoliaSearchKeys={algoliaSearchKeys}
+            analyticsStream={jest.fn()}
           >
             <ArticleExtras
               analyticsStream={() => {}}
@@ -135,6 +125,7 @@ export default () => {
           <AlgoliaSearchProvider
             article={{ id: "dummy-article-id" }}
             algoliaSearchKeys={algoliaSearchKeys}
+            analyticsStream={jest.fn()}
           >
             <ArticleExtras
               analyticsStream={() => {}}