+++ packages/article-in-depth/src/article-in-depth.web.js
@@ -77,7 +77,9 @@ class ArticlePage extends Component {
       isPreview,
       additionalRelatedArticlesFlag,
       algoliaSearchKeys,
-      inArticlePuffFlag
+      inArticlePuffFlag,
+      latestFromSectionFlag,
+      latestFromSection
     } = this.props;
 
     if (error || isLoading) {
@@ -97,6 +99,8 @@ class ArticlePage extends Component {
         additionalRelatedArticlesFlag={additionalRelatedArticlesFlag}
         algoliaSearchKeys={algoliaSearchKeys}
         inArticlePuffFlag={inArticlePuffFlag}
+        latestFromSectionFlag={latestFromSectionFlag}
+        latestFromSection={latestFromSection}
       />
     );
   }