+++ packages/article-magazine-standard/src/article-magazine-standard.web.js
@@ -70,7 +70,9 @@ class ArticlePage extends Component {
       isPreview,
       additionalRelatedArticlesFlag,
       algoliaSearchKeys,
-      inArticlePuffFlag
+      inArticlePuffFlag,
+      latestFromSectionFlag,
+      latestFromSection
     } = this.props;
 
     if (error || isLoading) {
@@ -91,6 +93,8 @@ class ArticlePage extends Component {
         additionalRelatedArticlesFlag={additionalRelatedArticlesFlag}
         algoliaSearchKeys={algoliaSearchKeys}
         inArticlePuffFlag={inArticlePuffFlag}
+        latestFromSectionFlag={latestFromSectionFlag}
+        latestFromSection={latestFromSection}
       />
     );
   }