+++ packages/ts-components/src/index.ts
@@ -23,12 +23,9 @@ export {
 } from './components/newsletter-puff/PreviewNewsletterPuff';
 
 export {
-  NewsletterPuffButton
-} from './components/newsletter-puff/NewsletterPuffButton';
-
-export {
-  NewsletterPuffLink
-} from './components/newsletter-puff/NewsletterPuffLink';
+  AlgoliaSearchProvider,
+  useAlgoliaSearch
+} from './helpers/algolia/AlgoliaSearchProvider';
 
 export * from './components/opta/football/fixtures/OptaFootballFixtures';
 export * from './components/opta/football/standings/OptaFootballStandings';
@@ -38,3 +35,7 @@ export * from './components/opta/football/match-stats/OptaFootballMatchStats';
 export {
   RelatedArticleSlice
 } from './components/related-article-slice/RelatedArticleSlice';
+
+export {
+  LatestFromSection
+} from './components/latest-from-section/LatestFromSection';