+++ packages/article-extras/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.11](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.10...@times-components/article-extras@0.7.11) (2021-07-07)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.7.10](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.9...@times-components/article-extras@0.7.10) (2021-07-05)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.7.9](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.8...@times-components/article-extras@0.7.9) (2021-07-02)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.7.8](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.7...@times-components/article-extras@0.7.8) (2021-06-29)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
 ## [0.7.7](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.7.6...@times-components/article-extras@0.7.7) (2021-06-23)
 
 **Note:** Version bump only for package @times-components/article-extras