+++ packages/pages/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.22](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.21...@times-components/pages@2.3.22) (2021-07-07)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.21](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.20...@times-components/pages@2.3.21) (2021-07-05)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.20](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.19...@times-components/pages@2.3.20) (2021-07-02)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.3.19](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.18...@times-components/pages@2.3.19) (2021-06-29)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
 ## [2.3.18](https://github.com/newsuk/times-components/compare/@times-components/pages@2.3.17...@times-components/pages@2.3.18) (2021-06-23)
 
 **Note:** Version bump only for package @times-components/pages